+++ example/lib/main.dart
@@ -1,16 +1,19 @@
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import 'package:introduction_screen/introduction_screen.dart';
-
-const kImageDemo =
-    "https://cdn4.iconfinder.com/data/icons/onboarding-material-color/128/__14-512.png";
 
 void main() => runApp(App());
 
 class App extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
+    SystemChrome.setSystemUIOverlayStyle(
+      SystemUiOverlayStyle.dark.copyWith(statusBarColor: Colors.transparent),
+    );
+
     return MaterialApp(
       title: 'Introduction screen',
+      debugShowCheckedModeBanner: false,
       theme: ThemeData(primarySwatch: Colors.blue),
       home: OnBoardingPage(),
     );
@@ -26,61 +29,73 @@ class OnBoardingPage extends StatelessWidget {
     );
   }
 
-  Widget _buildImage() {
+  Widget _buildImage(String assetName) {
     return Align(
-      child: Image.network(kImageDemo, height: 175.0),
+      child: Image.asset('assets/$assetName.jpg', width: 350.0),
       alignment: Alignment.bottomCenter,
     );
   }
 
   @override
   Widget build(BuildContext context) {
+    const bodyStyle = TextStyle(fontSize: 19.0);
+    const pageDecoration = const PageDecoration(
+      titleTextStyle: TextStyle(fontSize: 28.0, fontWeight: FontWeight.w700),
+      bodyTextStyle: bodyStyle,
+      descriptionPadding: EdgeInsets.fromLTRB(16.0, 0.0, 16.0, 16.0),
+      pageColor: Colors.white,
+      imagePadding: EdgeInsets.zero,
+    );
+
     return IntroductionScreen(
       pages: [
         PageViewModel(
-          title: "First title page",
-          body: "Text of the first page of this onboarding",
-          image: _buildImage(),
+          title: "Fractional shares",
+          body:
+              "Instead of having to buy an entire share, invest any amount you want.",
+          image: _buildImage('img1'),
+          decoration: pageDecoration,
         ),
         PageViewModel(
-          title: "Second title page",
-          body: "Text of the second page of this onboarding",
-          image: _buildImage(),
+          title: "Learn as you go",
+          body:
+              "Download the Stockpile app and master the market with our mini-lesson.",
+          image: _buildImage('img2'),
+          decoration: pageDecoration,
+        ),
+        PageViewModel(
+          title: "Kids and teens",
+          body:
+              "Kids and teens can track their stocks 24/7 and place trades that you approve.",
+          image: _buildImage('img3'),
+          decoration: pageDecoration,
+        ),
+        PageViewModel(
+          title: "Another title page",
+          body: "Another beautiful body text for this example onboarding",
+          image: _buildImage('img2'),
           footer: RaisedButton(
             onPressed: () {/* Nothing */},
-            child: const Text('Button', style: TextStyle(color: Colors.white)),
+            child: const Text('FooButton', style: TextStyle(color: Colors.white)),
             color: Colors.lightBlue,
+            shape: RoundedRectangleBorder(
+              borderRadius: BorderRadius.circular(8.0),
+            ),
           ),
+          decoration: pageDecoration,
         ),
         PageViewModel(
-          title: "Third title page",
-          body: "Text of the third page of this onboarding",
-          image: _buildImage(),
-          decoration: PageDecoration(
-            titleTextStyle: const TextStyle(
-              fontSize: 28.0,
-              fontWeight: FontWeight.w600,
-              color: Colors.red,
-            ),
-            bodyTextStyle: const TextStyle(fontSize: 22.0),
-            dotsDecorator: const DotsDecorator(
-              activeColor: Colors.red,
-              activeSize: Size.fromRadius(8),
-            ),
-            pageColor: Colors.grey[200],
-          ),
-        ),
-        PageViewModel(
-          title: "Fourth title page",
+          title: "Title of last page",
           bodyWidget: Row(
             mainAxisAlignment: MainAxisAlignment.center,
             children: const [
-              Text("Click on "),
+              Text("Click on ", style: bodyStyle),
               Icon(Icons.edit),
-              Text(" to edit a post"),
+              Text(" to edit a post", style: bodyStyle),
             ],
           ),
-          image: _buildImage(),
+          image: _buildImage('img1'),
+          decoration: pageDecoration,
         ),
       ],
       onDone: () => _onIntroEnd(context),
@@ -91,6 +106,14 @@ class OnBoardingPage extends StatelessWidget {
       skip: const Text('Skip'),
       next: const Icon(Icons.arrow_forward),
       done: const Text('Done', style: TextStyle(fontWeight: FontWeight.w600)),
+      dotsDecorator: const DotsDecorator(
+        size: Size(10.0, 10.0),
+        color: Color(0xFFBDBDBD),
+        activeSize: Size(22.0, 10.0),
+        activeShape: RoundedRectangleBorder(
+          borderRadius: BorderRadius.all(Radius.circular(25.0)),
+        ),
+      ),
     );
   }
 }