+++ lib/src/model/page_decoration.dart
@@ -5,9 +5,6 @@ class PageDecoration {
   /// Background page color
   final Color pageColor;
 
-  /// Dots decorator to custom dots color, size and spacing
-  final DotsDecorator dotsDecorator;
-
   /// TextStyle for title
   ///
   /// @Default style `fontSize: 20.0, fontWeight: FontWeight.bold`
@@ -27,9 +24,33 @@ class PageDecoration {
   /// Flex ratio of the body
   final int bodyFlex;
 
+  /// Padding of image
+  /// 
+  /// @Default `EdgeInsets.only(bottom: 24.0)`
+  final EdgeInsets imagePadding;
+
+  /// Padding of content (title + description + footer)
+  /// 
+  /// @Default `EdgeInsets.all(16.0)`
+  final EdgeInsets contentPadding;
+
+  /// Padding of title
+  /// 
+  /// @Default `EdgeInsets.only(bottom: 24.0)`
+  final EdgeInsets titlePadding;
+
+  /// Padding of description
+  /// 
+  /// @Default: `EdgeInsets.zero`
+  final EdgeInsets descriptionPadding;
+
+  /// Padding of footer
+  /// 
+  /// @Default `EdgeInsets.symmetric(vertical: 24.0)`
+  final EdgeInsets footerPadding;
+
   const PageDecoration({
     this.pageColor,
-    this.dotsDecorator = const DotsDecorator(),
     this.titleTextStyle = const TextStyle(
       fontSize: 20.0,
       fontWeight: FontWeight.bold,
@@ -41,6 +62,11 @@ class PageDecoration {
     this.boxDecoration,
     this.imageFlex = 1,
     this.bodyFlex = 1,
+    this.imagePadding = const EdgeInsets.only(bottom: 24.0),
+    this.contentPadding = const EdgeInsets.all(16.0),
+    this.titlePadding = const EdgeInsets.only(bottom: 24.0),
+    this.descriptionPadding = EdgeInsets.zero,
+    this.footerPadding = const EdgeInsets.symmetric(vertical: 24.0),
   }) : assert(pageColor == null || boxDecoration == null,
             'Cannot provide both a Color and a BoxDecoration\n');
 }